Validate urls for further CSRF protection
